To Create Domain Free Instance Sample 01: Organize Classes To Fulfill Their Obligations
Rename a class DomainFreeDictionary to CrossDomainDictionary because it isn't so free. The keys and values must be to extend MarshalByRefObject or to set SerializableAttribute, if  you want to set them to the dictionary.
Rename a class DomainFreeInstantiator to DomainLooser because it isn't so free, too. The instantiator must be a reference of a static method.